+++ translations/holyscreen_pt_BR.ts
@@ -1006,57 +1006,62 @@
         <translation>Pesquisar mídia...</translation>
     </message>
     <message>
-        <location filename="../src/ui/operator/LibraryPanel.qml" line="116"/>
+        <location filename="../src/ui/operator/LibraryPanel.qml" line="93"/>
+        <source>Pesquisar mídia</source>
+        <translation>Pesquisar mídia</translation>
+    </message>
+    <message>
+        <location filename="../src/ui/operator/LibraryPanel.qml" line="119"/>
         <source>Letras</source>
         <translation>Letras</translation>
     </message>
     <message>
-        <location filename="../src/ui/operator/LibraryPanel.qml" line="116"/>
+        <location filename="../src/ui/operator/LibraryPanel.qml" line="119"/>
         <source>Áudio</source>
         <translation>Áudio</translation>
     </message>
     <message>
-        <location filename="../src/ui/operator/LibraryPanel.qml" line="117"/>
+        <location filename="../src/ui/operator/LibraryPanel.qml" line="120"/>
         <source>Vídeo</source>
         <translation>Vídeo</translation>
     </message>
     <message>
-        <location filename="../src/ui/operator/LibraryPanel.qml" line="117"/>
+        <location filename="../src/ui/operator/LibraryPanel.qml" line="120"/>
         <source>Imagem</source>
         <translation>Imagem</translation>
     </message>
     <message>
-        <location filename="../src/ui/operator/LibraryPanel.qml" line="170"/>
+        <location filename="../src/ui/operator/LibraryPanel.qml" line="173"/>
         <source>Sem título</source>
         <translation>Sem título</translation>
     </message>
     <message>
-        <location filename="../src/ui/operator/LibraryPanel.qml" line="177"/>
+        <location filename="../src/ui/operator/LibraryPanel.qml" line="180"/>
         <source>Letra</source>
         <translation>Letra</translation>
     </message>
     <message>
-        <location filename="../src/ui/operator/LibraryPanel.qml" line="186"/>
+        <location filename="../src/ui/operator/LibraryPanel.qml" line="189"/>
         <source>ABRIR</source>
         <translation>ABRIR</translation>
     </message>
     <message>
-        <location filename="../src/ui/operator/LibraryPanel.qml" line="208"/>
+        <location filename="../src/ui/operator/LibraryPanel.qml" line="211"/>
         <source>Adicione pastas na Biblioteca para ver suas mídias</source>
         <translation>Adicione pastas na Biblioteca para ver suas mídias</translation>
     </message>
     <message>
-        <location filename="../src/ui/operator/LibraryPanel.qml" line="209"/>
+        <location filename="../src/ui/operator/LibraryPanel.qml" line="212"/>
         <source>Nenhum arquivo encontrado</source>
         <translation>Nenhum arquivo encontrado</translation>
     </message>
     <message>
-        <location filename="../src/ui/operator/LibraryPanel.qml" line="222"/>
+        <location filename="../src/ui/operator/LibraryPanel.qml" line="225"/>
         <source>%1 itens</source>
         <translation>%1 itens</translation>
     </message>
     <message>
-        <location filename="../src/ui/operator/LibraryPanel.qml" line="228"/>
+        <location filename="../src/ui/operator/LibraryPanel.qml" line="231"/>
         <source>Biblioteca</source>
         <translation>Biblioteca</translation>
     </message>